remove `Zenstruck\Collection\Doctrine\ORM\Repository`
