+++ src/Collection/Doctrine/ORM/Repository/IsCollection.php
@@ -3,11 +3,11 @@
 namespace Zenstruck\Collection\Doctrine\ORM\Repository;
 
 use Zenstruck\Collection;
-use Zenstruck\Collection\Doctrine\ORM\Repository;
+use Zenstruck\Collection\Doctrine\ORM\ObjectRepository;
 
 /**
- * Enables your repository to implement Zenstruck\Collection (and use
- * the Zenstruck\Collection\Paginatable trait).
+ * Enables your repository to implement {@see Collection} (and use
+ * the {@see Collection\Paginatable} trait).
  *
  * @author Kevin Bond <kevinbond@gmail.com>
  *
@@ -20,8 +20,8 @@ trait IsCollection
      */
     final public function take(int $limit, int $offset = 0): Collection
     {
-        if (!$this instanceof Repository) {
-            throw new \BadMethodCallException(\sprintf('"%s" can only be used on instances of "%s".', __TRAIT__, Repository::class));
+        if (!$this instanceof ObjectRepository) {
+            throw new \BadMethodCallException(\sprintf('"%s" can only be used on instances of "%s".', __TRAIT__, ObjectRepository::class));
         }
 
         return static::createResult($this->qb())->take($limit, $offset);
